requirements.txt
aiocoap==0.4.14
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
amqp==5.3.1
anyio==1.4.0
asgiref==3.8.1
async-generator==1.10
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
celery-stubs==0.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.1
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-debug-toolbar==5.2.0
django-jalali==7.4.0
django-jalali-date==2.0.0
django-js-asset==3.1.2
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-timezone-field==7.1
djangorestframework==3.16.0
dotenv==0.9.9
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.4
MarkupSafe==3.0.2
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
pytest-mock==3.14.1
pytest-random-order==1.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
setuptools==80.9.0
sitemaps==0.1.0
sites==0.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
